[refactor] rename DefaultValues to Schemas and move to User folder
