+++ App.js
@@ -2,7 +2,7 @@ import React from 'react';
 // change to import RootComp from '../<expo-sample-app-folder>/Framework/RootComp';
 import RootComp from './Framework/RootComp';
 import { screenMaps, DEFAULT_SCREEN } from './User/PageMapper';
-import { NEW_USER_DATA, APP_NAME } from './DefaultValues';
+import { NEW_USER_DATA, APP_NAME } from './User/Schemas';
 
 const App = () => {
   return (
